+++ cedar-policy/src/api.rs
@@ -23,6 +23,7 @@
 pub use ast::Effect;
 pub use authorizer::Decision;
 use cedar_policy_core::ast;
+use cedar_policy_core::ast::RestrictedExprError;
 use cedar_policy_core::authorizer;
 use cedar_policy_core::entities;
 use cedar_policy_core::entities::JsonDeserializationErrorContext;
@@ -35,8 +36,9 @@ use cedar_policy_core::parser;
 pub use cedar_policy_core::parser::err::ParseErrors;
 use cedar_policy_core::parser::SourceInfo;
 use cedar_policy_core::FromNormalizedStr;
-pub use cedar_policy_validator::{TypeErrorKind, ValidationErrorKind, ValidationWarningKind};
-use itertools::Itertools;
+pub use cedar_policy_validator::{
+    TypeErrorKind, UnsupportedFeature, ValidationErrorKind, ValidationWarningKind,
+};
 use ref_cast::RefCast;
 use serde::{Deserialize, Serialize};
 use smol_str::SmolStr;
@@ -169,6 +171,15 @@ impl Entities {
         Self(self.0.partial())
     }
 
+    /// Attempt to eagerly compute the values of attributes for all entities in the slice.
+    /// This can fail if evaluation of the [`RestrictedExpression`] fails.
+    /// In a future major version, we will likely make this function automatically called via the constructor.
+    pub fn evaluate(self) -> Result<Self, EvaluationError> {
+        Ok(Self(self.0.evaluate().map_err(|e| {
+            EvaluationError::StringMessage(e.to_string())
+        })?))
+    }
+
     /// Iterate over the `Entity`'s in the `Entities`
     pub fn iter(&self) -> impl Iterator<Item = &Entity> {
         self.0.iter().map(Entity::ref_cast)
@@ -289,6 +300,53 @@ impl Default for Authorizer {
 
 impl Authorizer {
     /// Create a new `Authorizer`
+    /// ```
+
+    /// # use cedar_policy::{Authorizer, Context, Entities, EntityId, EntityTypeName,
+    /// # EntityUid, Request,PolicySet};
+    /// # use std::str::FromStr;
+    /// # // create a request
+    /// # let p_eid = EntityId::from_str("alice").unwrap();
+    /// # let p_name: EntityTypeName = EntityTypeName::from_str("User").unwrap();
+    /// # let p = EntityUid::from_type_name_and_id(p_name, p_eid);
+    /// #
+    /// # let a_eid = EntityId::from_str("view").unwrap();
+    /// # let a_name: EntityTypeName = EntityTypeName::from_str("Action").unwrap();
+    /// # let a = EntityUid::from_type_name_and_id(a_name, a_eid);
+    /// #
+    /// # let r_eid = EntityId::from_str("trip").unwrap();
+    /// # let r_name: EntityTypeName = EntityTypeName::from_str("Album").unwrap();
+    /// # let r = EntityUid::from_type_name_and_id(r_name, r_eid);
+    /// #
+    /// # let c = Context::empty();
+    /// #
+    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), c);
+    /// #
+    /// # // create a policy
+    /// # let s = r#"permit(
+    /// #     principal == User::"alice",
+    /// #     action == Action::"view",
+    /// #     resource == Album::"trip"
+    /// #   )when{
+    /// #     principal.ip_addr.isIpv4()
+    /// #   };
+    /// # "#;
+    /// # let policy = PolicySet::from_str(s).expect("policy error");
+    /// # // create entities
+    /// # let e = r#"[
+    /// #     {
+    /// #         "uid": {"type":"User","id":"alice"},
+    /// #         "attrs": {
+    /// #             "age":19,
+    /// #             "ip_addr":{"__extn":{"fn":"ip", "arg":"10.0.1.101"}}
+    /// #         },
+    /// #         "parents": []
+    /// #     }
+    /// # ]"#;
+    /// # let entities = Entities::from_json_str(e, None).expect("entity error");
+    /// let authorizer = Authorizer::new();
+    /// let r = authorizer.is_authorized(&request, &policy, &entities);
+    /// ```
     pub fn new() -> Self {
         Self(authorizer::Authorizer::new())
     }
@@ -298,6 +356,56 @@ impl Authorizer {
     ///
     /// The language spec and Dafny model give a precise definition of how this
     /// is computed.
+    /// ```
+    /// use cedar_policy::{Authorizer,Context,Entities,EntityId,EntityTypeName,
+    /// EntityUid, Request,PolicySet};
+    /// use std::str::FromStr;
+    ///
+    /// // create a request
+    /// let p_eid = EntityId::from_str("alice").unwrap();
+    /// let p_name: EntityTypeName = EntityTypeName::from_str("User").unwrap();
+    /// let p = EntityUid::from_type_name_and_id(p_name, p_eid);
+    ///
+    /// let a_eid = EntityId::from_str("view").unwrap();
+    /// let a_name: EntityTypeName = EntityTypeName::from_str("Action").unwrap();
+    /// let a = EntityUid::from_type_name_and_id(a_name, a_eid);
+    ///
+    /// let r_eid = EntityId::from_str("trip").unwrap();
+    /// let r_name: EntityTypeName = EntityTypeName::from_str("Album").unwrap();
+    /// let r = EntityUid::from_type_name_and_id(r_name, r_eid);
+    ///
+    /// let c = Context::empty();
+    ///
+    /// let request: Request = Request::new(Some(p), Some(a), Some(r), c);
+    ///
+    /// // create a policy
+    /// let s = r#"
+    /// permit (
+    ///   principal == User::"alice",
+    ///   action == Action::"view",
+    ///   resource == Album::"trip"
+    /// )
+    /// when { principal.ip_addr.isIpv4() };
+    /// "#;
+    /// let policy = PolicySet::from_str(s).expect("policy error");
+
+    /// // create entities
+    /// let e = r#"[
+    ///     {
+    ///         "uid": {"type":"User","id":"alice"},
+    ///         "attrs": {
+    ///             "age":19,
+    ///             "ip_addr":{"__extn":{"fn":"ip", "arg":"10.0.1.101"}}
+    ///         },
+    ///         "parents": []
+    ///     }
+    /// ]"#;
+    /// let entities = Entities::from_json_str(e, None).expect("entity error");
+    ///
+    /// let authorizer = Authorizer::new();
+    /// let r = authorizer.is_authorized(&request, &policy, &entities);
+    /// println!("{:?}", r);
+    /// ```
     pub fn is_authorized(&self, r: &Request, p: &PolicySet, e: &Entities) -> Response {
         self.0.is_authorized(&r.0, &p.ast, &e.0).into()
     }
@@ -631,80 +739,77 @@ impl Schema {
 /// Errors encountered during construction of a Validation Schema
 #[derive(Debug, Error)]
 pub enum SchemaError {
-    /// Errors loading and parsing schema files
-    #[error("JSON Schema file could not be parsed: {0}")]
-    ParseJson(serde_json::Error),
+    /// Error thrown by the `serde_json` crate during deserialization
+    #[error("failed to parse schema: {0}")]
+    Serde(#[from] serde_json::Error),
     /// Errors occurring while computing or enforcing transitive closure on
-    /// action id hierarchy.
-    #[error("Transitive closure error on action hierarchy: {0}")]
-    ActionTransitiveClosureError(String),
+    /// action hierarchy.
+    #[error("transitive closure computation/enforcement error on action hierarchy: {0}")]
+    ActionTransitiveClosure(String),
     /// Errors occurring while computing or enforcing transitive closure on
     /// entity type hierarchy.
-    #[error("Transitive closure error on entity hierarchy: {0}")]
-    EntityTransitiveClosureError(String),
-    /// Error generated when processing a schema file that uses features which
-    /// are not yet supported by the implementation.
-    #[error("Unsupported feature used in schema: {0}")]
-    UnsupportedSchemaFeature(String),
-    /// Undeclared entity type(s) used in an entity type's memberOf field or an
-    /// action's appliesTo fields.
-    #[error("Undeclared entity types: {0:?}")]
+    #[error("transitive closure computation/enforcement error on entity type hierarchy: {0}")]
+    EntityTypeTransitiveClosure(String),
+    /// Error generated when processing a schema file that uses unsupported features
+    #[error("unsupported feature used in schema: {0}")]
+    UnsupportedFeature(String),
+    /// Undeclared entity type(s) used in the `memberOf` field of an entity
+    /// type, the `appliesTo` fields of an action, or an attribute type in a
+    /// context or entity attribute record. Entity types in the error message
+    /// are fully qualified, including any implicit or explicit namespaces.
+    #[error("undeclared entity type(s): {0:?}")]
     UndeclaredEntityTypes(HashSet<String>),
-    /// Undeclared action(s) used in an action's memberOf field.
-    #[error("Undeclared actions: {0:?}")]
+    /// Undeclared action(s) used in the `memberOf` field of an action.
+    #[error("undeclared action(s): {0:?}")]
     UndeclaredActions(HashSet<String>),
-    /// Undeclared type used in entity or context attributes.
-    #[error("Undeclared common types: {0:?}")]
-    UndeclaredCommonType(HashSet<String>),
+    /// Undeclared common type(s) used in entity or context attributes.
+    #[error("undeclared common type(s): {0:?}")]
+    UndeclaredCommonTypes(HashSet<String>),
     /// Duplicate specifications for an entity type. Argument is the name of
     /// the duplicate entity type.
-    #[error("Duplicate entity type {0}")]
+    #[error("duplicate entity type: {0}")]
     DuplicateEntityType(String),
     /// Duplicate specifications for an action. Argument is the name of the
     /// duplicate action.
-    #[error("Duplicate action {0}")]
+    #[error("duplicate action: {0}")]
     DuplicateAction(String),
-    /// Duplicate specifications for a reusable common type. Argument is the
-    /// name of the duplicate type.
-    #[error("Duplicate common type {0}")]
+    /// Duplicate specification for a reusable type declaration.
+    #[error("duplicate common type: {0}")]
     DuplicateCommonType(String),
     /// Cycle in the schema's action hierarchy.
-    #[error("Cycle in action hierarchy")]
+    #[error("cycle in action hierarchy")]
     CycleInActionHierarchy,
     /// Parse errors occurring while parsing an entity type.
-    #[error("Parse error in entity type: {0}")]
-    EntityTypeParse(ParseErrors),
+    #[error("parse error in entity type: {0}")]
+    ParseEntityType(ParseErrors),
     /// Parse errors occurring while parsing a namespace identifier.
-    #[error("Parse error in namespace identifier: {0}")]
-    NamespaceParse(ParseErrors),
-    /// Parse errors occurring while parsing a common type identifier.
-    #[error("Parse error in common type identifier: {0}")]
-    CommonTypeParseError(ParseErrors),
+    #[error("parse error in namespace identifier: {0}")]
+    ParseNamespace(ParseErrors),
     /// Parse errors occurring while parsing an extension type.
-    #[error("Parse error in extension type: {0}")]
-    ExtensionTypeParse(ParseErrors),
+    #[error("parse error in extension type: {0}")]
+    ParseExtensionType(ParseErrors),
+    /// Parse errors occurring while parsing the name of a reusable
+    /// declared type.
+    #[error("parse error in common type identifier: {0}")]
+    ParseCommonType(ParseErrors),
     /// The schema file included an entity type `Action` in the entity type
     /// list. The `Action` entity type is always implicitly declared, and it
     /// cannot currently have attributes or be in any groups, so there is no
     /// purposes in adding an explicit entry.
-    #[error("Entity type `Action` declared in `entityTypes` list.")]
+    #[error("entity type `Action` declared in `entityTypes` list")]
     ActionEntityTypeDeclared,
-    /// One or more action entities are declared with `attributes` lists, but
-    /// action entities cannot have attributes.
-    #[error("Actions declared with `attributes`: [{}]", .0.iter().map(String::as_str).join(", "))]
-    ActionEntityAttributes(Vec<String>),
-    /// An action context or entity type shape was declared to have a type other
-    /// than `Record`.
-    #[error("{0} is not a record")]
+    /// `context` or `shape` fields are not records
+    #[error("`{0}` is not a record")]
     ContextOrShapeNotRecord(ContextOrShape),
-    /// An Action Entity (transitively) has an attribute that is an empty set
-    #[error("Action attribute is an empty set")]
-    ActionEntityAttributeEmptySet,
-    /// An Action Entity (transitively) has an attribute of unsupported type (ExprEscape, EntityEscape or ExtnEscape)
-    #[error(
-        "Action has an attribute of unsupported type (escaped expression, entity or extension)"
-    )]
-    ActionEntityAttributeUnsupportedType,
+    /// An action entity (transitively) has an attribute that is an empty set.
+    /// The validator cannot assign a type to an empty set.
+    /// This error variant should only be used when `PermitAttributes` is enabled.
+    #[error("action `{0}` has an attribute that is an empty set")]
+    ActionAttributesContainEmptySet(EntityUid),
+    /// An action entity (transitively) has an attribute of unsupported type (`ExprEscape`, `EntityEscape` or `ExtnEscape`).
+    /// This error variant should only be used when `PermitAttributes` is enabled.
+    #[error("action `{0}` has an attribute with unsupported JSON representation: {1}")]
+    UnsupportedActionAttribute(EntityUid, String),
 }
 
 /// Describes in what action context or entity type shape a schema parsing error
@@ -747,22 +852,22 @@ impl From<cedar_policy_validator::ContextOrShape> for ContextOrShape {
 impl From<cedar_policy_validator::SchemaError> for SchemaError {
     fn from(value: cedar_policy_validator::SchemaError) -> Self {
         match value {
-            cedar_policy_validator::SchemaError::ParseFileFormat(e) => Self::ParseJson(e),
-            cedar_policy_validator::SchemaError::ActionTransitiveClosureError(e) => {
-                Self::ActionTransitiveClosureError(e.to_string())
+            cedar_policy_validator::SchemaError::Serde(e) => Self::Serde(e),
+            cedar_policy_validator::SchemaError::ActionTransitiveClosure(e) => {
+                Self::ActionTransitiveClosure(e.to_string())
             }
-            cedar_policy_validator::SchemaError::EntityTransitiveClosureError(e) => {
-                Self::EntityTransitiveClosureError(e.to_string())
+            cedar_policy_validator::SchemaError::EntityTypeTransitiveClosure(e) => {
+                Self::EntityTypeTransitiveClosure(e.to_string())
             }
-            cedar_policy_validator::SchemaError::UnsupportedSchemaFeature(e) => {
-                Self::UnsupportedSchemaFeature(e.to_string())
+            cedar_policy_validator::SchemaError::UnsupportedFeature(e) => {
+                Self::UnsupportedFeature(e.to_string())
             }
             cedar_policy_validator::SchemaError::UndeclaredEntityTypes(e) => {
                 Self::UndeclaredEntityTypes(e)
             }
             cedar_policy_validator::SchemaError::UndeclaredActions(e) => Self::UndeclaredActions(e),
-            cedar_policy_validator::SchemaError::UndeclaredCommonType(c) => {
-                Self::UndeclaredCommonType(c)
+            cedar_policy_validator::SchemaError::UndeclaredCommonTypes(c) => {
+                Self::UndeclaredCommonTypes(c)
             }
             cedar_policy_validator::SchemaError::DuplicateEntityType(e) => {
                 Self::DuplicateEntityType(e)
@@ -774,30 +879,23 @@ impl From<cedar_policy_validator::SchemaError> for SchemaError {
             cedar_policy_validator::SchemaError::CycleInActionHierarchy => {
                 Self::CycleInActionHierarchy
             }
-            cedar_policy_validator::SchemaError::EntityTypeParseError(e) => {
-                Self::EntityTypeParse(e)
-            }
-            cedar_policy_validator::SchemaError::NamespaceParseError(e) => Self::NamespaceParse(e),
-            cedar_policy_validator::SchemaError::CommonTypeParseError(e) => {
-                Self::CommonTypeParseError(e)
-            }
-            cedar_policy_validator::SchemaError::ExtensionTypeParseError(e) => {
-                Self::ExtensionTypeParse(e)
+            cedar_policy_validator::SchemaError::ParseEntityType(e) => Self::ParseEntityType(e),
+            cedar_policy_validator::SchemaError::ParseNamespace(e) => Self::ParseNamespace(e),
+            cedar_policy_validator::SchemaError::ParseCommonType(e) => Self::ParseCommonType(e),
+            cedar_policy_validator::SchemaError::ParseExtensionType(e) => {
+                Self::ParseExtensionType(e)
             }
             cedar_policy_validator::SchemaError::ActionEntityTypeDeclared => {
                 Self::ActionEntityTypeDeclared
             }
-            cedar_policy_validator::SchemaError::ActionEntityAttributes(e) => {
-                Self::ActionEntityAttributes(e)
-            }
             cedar_policy_validator::SchemaError::ContextOrShapeNotRecord(context_or_shape) => {
                 Self::ContextOrShapeNotRecord(context_or_shape.into())
             }
-            cedar_policy_validator::SchemaError::ActionEntityAttributeEmptySet => {
-                Self::ActionEntityAttributeEmptySet
+            cedar_policy_validator::SchemaError::ActionAttributesContainEmptySet(uid) => {
+                Self::ActionAttributesContainEmptySet(EntityUid(uid))
             }
-            cedar_policy_validator::SchemaError::ActionEntityAttributeUnsupportedType => {
-                Self::ActionEntityAttributeUnsupportedType
+            cedar_policy_validator::SchemaError::UnsupportedActionAttribute(uid, escape_type) => {
+                Self::UnsupportedActionAttribute(EntityUid(uid), escape_type)
             }
         }
     }
@@ -1184,12 +1282,16 @@ impl FromStr for PolicySet {
     /// See [`Policy`] for more.
     fn from_str(policies: &str) -> Result<Self, Self::Err> {
         let (texts, pset) = parser::parse_policyset_and_also_return_policy_text(policies)?;
+        // PANIC SAFETY: By the invariant on `parse_policyset_and_also_return_policy_text(policies)`, every `PolicyId` in `pset.policies()` occurs as a key in `text`.
+        #[allow(clippy::expect_used)]
         let policies = pset.policies().map(|p|
             (
                 PolicyId(p.id().clone()),
                 Policy { lossless: LosslessPolicy::policy_or_template_text(*texts.get(p.id()).expect("internal invariant violation: policy id exists in asts but not texts")), ast: p.clone() }
             )
         ).collect();
+        // PANIC SAFETY: By the same invariant, every `PolicyId` in `pset.templates()` also occurs as a key in `text`.
+        #[allow(clippy::expect_used)]
         let templates = pset.templates().map(|t|
             (
                 PolicyId(t.id().clone()),
@@ -1321,6 +1423,8 @@ impl PolicySet {
                 unwrapped_vals.clone(),
             )
             .map_err(PolicySetError::LinkingError)?;
+        // PANIC SAFETY: `lossless.link()` will not fail after `ast.link()` succeeds
+        #[allow(clippy::expect_used)]
         let linked_lossless = self
             .templates
             .get(&template_id)
@@ -1771,6 +1875,12 @@ impl Policy {
         }
     }
 
+    /// To avoid panicking, this function may only be called when `slot` is the
+    /// SlotId corresponding to the scope constraint from which the entity
+    /// reference `r` was extracted. I.e., If `r` is taken from the principal
+    /// scope constraint, `slot` must be `?principal`. This ensures that the
+    /// SlotId exists in the policy (and therefore the slot environment map)
+    /// whenever the EntityReference `r` is the Slot variant.
     fn convert_entity_reference<'a>(
         &'a self,
         r: &'a ast::EntityReference,
@@ -1778,7 +1888,8 @@ impl Policy {
     ) -> &'a EntityUid {
         match r {
             ast::EntityReference::EUID(euid) => EntityUid::ref_cast(euid),
-            // This `unwrap` here is safe due the invariant (values total map) on policies.
+            // PANIC SAFETY: This `unwrap` here is safe due the invariant (values total map) on policies.
+            #[allow(clippy::unwrap_used)]
             ast::EntityReference::Slot => EntityUid::ref_cast(self.ast.env().get(&slot).unwrap()),
         }
     }
@@ -1800,6 +1911,68 @@ impl Policy {
     /// If `id` is Some, the policy will be given that Policy Id.
     /// If `id` is None, then "JSON policy" will be used.
     /// The behavior around None may change in the future.
+    ///
+    /// ```
+    /// use cedar_policy::{Policy, PolicyId};
+    /// use std::str::FromStr;
+    ///
+    /// let data : serde_json::Value = serde_json::json!(
+    ///        {
+    ///            "effect":"permit",
+    ///            "principal":{
+    ///            "op":"==",
+    ///            "entity":{
+    ///                "type":"User",
+    ///                "id":"bob"
+    ///            }
+    ///            },
+    ///            "action":{
+    ///            "op":"==",
+    ///            "entity":{
+    ///                "type":"Action",
+    ///                "id":"view"
+    ///            }
+    ///            },
+    ///            "resource":{
+    ///            "op":"==",
+    ///            "entity":{
+    ///                "type":"Album",
+    ///                "id":"trip"
+    ///            }
+    ///            },
+    ///            "conditions":[
+    ///            {
+    ///                "kind":"when",
+    ///                "body":{
+    ///                   ">":{
+    ///                        "left":{
+    ///                        ".":{
+    ///                            "left":{
+    ///                                "Var":"principal"
+    ///                            },
+    ///                            "attr":"age"
+    ///                        }
+    ///                        },
+    ///                        "right":{
+    ///                        "Value":18
+    ///                        }
+    ///                    }
+    ///                }
+    ///            }
+    ///            ]
+    ///        }
+    /// );
+    /// let policy = Policy::from_json(None, data).unwrap();
+    /// let src = r#"
+    ///   permit(
+    ///     principal == User::"bob",
+    ///     action == Action::"view",
+    ///     resource == Album::"trip"
+    ///   )
+    ///   when { principal.age > 18 };"#;
+    /// let expected_output = Policy::parse(None, src).unwrap();
+    /// assert_eq!(policy.to_string(), expected_output.to_string());
+    /// ```
     pub fn from_json(
         id: Option<PolicyId>,
         json: serde_json::Value,
@@ -1813,6 +1986,23 @@ impl Policy {
     }
 
     /// Get the JSON representation of this `Policy`.
+    ///  ```
+    /// use cedar_policy::Policy;
+    /// let src = r#"
+    ///   permit(
+    ///     principal == User::"bob",
+    ///     action == Action::"view",
+    ///     resource == Album::"trip"
+    ///   )
+    ///   when { principal.age > 18 };"#;
+
+    /// let policy = Policy::parse(None, src).unwrap();
+    /// println!("{}", policy);
+    /// // convert the policy to JSON
+    /// let json = policy.to_json().unwrap();
+    /// println!("{}", json);
+    /// assert_eq!(policy.to_string(), Policy::from_json(None, json).unwrap().to_string());
+    /// ```
     pub fn to_json(&self) -> Result<serde_json::Value, impl std::error::Error> {
         let est = self.lossless.est()?;
         let json = serde_json::to_value(est)?;
@@ -2025,11 +2215,98 @@ impl RestrictedExpression {
 }
 
 impl FromStr for RestrictedExpression {
-    type Err = ParseErrors;
+    type Err = RestrictedExprError;
 
     /// create a `RestrictedExpression` using Cedar syntax
     fn from_str(expression: &str) -> Result<Self, Self::Err> {
         ast::RestrictedExpr::from_str(expression).map(RestrictedExpression)
+    }
+}
+
+/// Builder for a [`Request`]
+///
+/// Note that you can create the `EntityUid`s using `.parse()` on any
+/// string (via the `FromStr` implementation for `EntityUid`).
+/// The principal, action, and resource fields are optional to support
+/// the case where these fields do not contribute to authorization
+/// decisions (e.g., because they are not used in your policies).
+/// If any of the fields are `None`, we will automatically generate
+/// a unique entity UID that is not equal to any UID in the store.
+///
+/// The default for principal, action and resource fields is Unknown.
+#[cfg(feature = "partial-eval")]
+#[derive(Debug, Default)]
+pub struct RequestBuilder {
+    principal: Option<ast::EntityUIDEntry>,
+    action: Option<ast::EntityUIDEntry>,
+    resource: Option<ast::EntityUIDEntry>,
+    context: Option<ast::Context>,
+}
+
+#[cfg(feature = "partial-eval")]
+impl RequestBuilder {
+    /// Set the principal
+    pub fn principal(self, principal: Option<EntityUid>) -> Self {
+        Self {
+            principal: Some(match principal {
+                Some(p) => ast::EntityUIDEntry::concrete(p.0),
+                None => ast::EntityUIDEntry::concrete(ast::EntityUID::unspecified_from_eid(
+                    ast::Eid::new("principal"),
+                )),
+            }),
+            ..self
+        }
+    }
+
+    /// Set the action
+    pub fn action(self, action: Option<EntityUid>) -> Self {
+        Self {
+            action: Some(match action {
+                Some(a) => ast::EntityUIDEntry::concrete(a.0),
+                None => ast::EntityUIDEntry::concrete(ast::EntityUID::unspecified_from_eid(
+                    ast::Eid::new("action"),
+                )),
+            }),
+            ..self
+        }
+    }
+
+    /// Set the resource
+    pub fn resource(self, resource: Option<EntityUid>) -> Self {
+        Self {
+            resource: Some(match resource {
+                Some(r) => ast::EntityUIDEntry::concrete(r.0),
+                None => ast::EntityUIDEntry::concrete(ast::EntityUID::unspecified_from_eid(
+                    ast::Eid::new("resource"),
+                )),
+            }),
+            ..self
+        }
+    }
+
+    /// Set the context
+    pub fn context(self, context: Context) -> Self {
+        Self {
+            context: Some(context.0),
+            ..self
+        }
+    }
+
+    /// Create the [`Request`]
+    pub fn build(self) -> Request {
+        let p = match self.principal {
+            Some(p) => p,
+            None => ast::EntityUIDEntry::Unknown,
+        };
+        let a = match self.action {
+            Some(a) => a,
+            None => ast::EntityUIDEntry::Unknown,
+        };
+        let r = match self.resource {
+            Some(r) => r,
+            None => ast::EntityUIDEntry::Unknown,
+        };
+        Request(ast::Request::new_with_unknowns(p, a, r, self.context))
     }
 }
 
@@ -2039,6 +2316,12 @@ impl FromStr for RestrictedExpression {
 pub struct Request(pub(crate) ast::Request);
 
 impl Request {
+    /// Create a [`RequestBuilder`]
+    #[cfg(feature = "partial-eval")]
+    pub fn builder() -> RequestBuilder {
+        RequestBuilder::default()
+    }
+
     /// Create a Request.
     ///
     /// Note that you can create the `EntityUid`s using `.parse()` on any
@@ -2101,6 +2384,11 @@ pub struct Context(ast::Context);
 
 impl Context {
     /// Create an empty `Context`
+    /// ```
+    /// use cedar_policy::Context;
+    /// let c = Context::empty();
+    /// // let request: Request = Request::new(Some(principal), Some(action), Some(resource), c);
+    /// ```
     pub fn empty() -> Self {
         Self(ast::Context::empty())
     }
@@ -2108,6 +2396,37 @@ impl Context {
     /// Create a `Context` from a map of key to "restricted expression",
     /// or a Vec of `(key, restricted expression)` pairs, or any other iterator
     /// of `(key, restricted expression)` pairs.
+    /// ```
+    /// use cedar_policy::{Context, RestrictedExpression};
+    /// use std::collections::HashMap;
+    /// use std::str::FromStr;
+    /// # use cedar_policy::{Entities, EntityId, EntityTypeName, EntityUid, Request,PolicySet};
+    /// let data : serde_json::Value = serde_json::json!({
+    ///     "sub": "1234",
+    ///     "groups": {
+    ///         "1234": {
+    ///             "group_id": "abcd",
+    ///             "group_name": "test-group"
+    ///         }
+    ///     }
+    /// });
+    /// let mut groups: HashMap<String, RestrictedExpression> = HashMap::new();
+    /// groups.insert("key".to_string(), RestrictedExpression::from_str(&data.to_string()).unwrap());
+    /// groups.insert("age".to_string(), RestrictedExpression::from_str("18").unwrap());
+    /// let context = Context::from_pairs(groups);
+    /// # // create a request
+    /// # let p_eid = EntityId::from_str("alice").unwrap();
+    /// # let p_name: EntityTypeName = EntityTypeName::from_str("User").unwrap();
+    /// # let p = EntityUid::from_type_name_and_id(p_name, p_eid);
+    /// #
+    /// # let a_eid = EntityId::from_str("view").unwrap();
+    /// # let a_name: EntityTypeName = EntityTypeName::from_str("Action").unwrap();
+    /// # let a = EntityUid::from_type_name_and_id(a_name, a_eid);
+    /// # let r_eid = EntityId::from_str("trip").unwrap();
+    /// # let r_name: EntityTypeName = EntityTypeName::from_str("Album").unwrap();
+    /// # let r = EntityUid::from_type_name_and_id(r_name, r_eid);
+    /// let request: Request = Request::new(Some(p), Some(a), Some(r), context);
+    /// ```
     pub fn from_pairs(pairs: impl IntoIterator<Item = (String, RestrictedExpression)>) -> Self {
         Self(ast::Context::from_pairs(
             pairs.into_iter().map(|(k, v)| (SmolStr::from(k), v.0)),
@@ -2124,6 +2443,34 @@ impl Context {
     /// if attributes have the wrong types (e.g., string instead of integer).
     /// Since different Actions have different schemas for `Context`, you also
     /// must specify the `Action` for schema-based parsing.
+    /// ```
+    /// use cedar_policy::{Context, RestrictedExpression};
+    /// use std::collections::HashMap;
+    /// use std::str::FromStr;
+    /// # use cedar_policy::{Entities, EntityId, EntityTypeName, EntityUid, Request,PolicySet};
+    /// let data =r#"{
+    ///     "sub": "1234",
+    ///     "groups": {
+    ///         "1234": {
+    ///             "group_id": "abcd",
+    ///             "group_name": "test-group"
+    ///         }
+    ///     }
+    /// }"#;
+    /// let context = Context::from_json_str(data, None).unwrap();
+    /// # // create a request
+    /// # let p_eid = EntityId::from_str("alice").unwrap();
+    /// # let p_name: EntityTypeName = EntityTypeName::from_str("User").unwrap();
+    /// # let p = EntityUid::from_type_name_and_id(p_name, p_eid);
+    /// #
+    /// # let a_eid = EntityId::from_str("view").unwrap();
+    /// # let a_name: EntityTypeName = EntityTypeName::from_str("Action").unwrap();
+    /// # let a = EntityUid::from_type_name_and_id(a_name, a_eid);
+    /// # let r_eid = EntityId::from_str("trip").unwrap();
+    /// # let r_name: EntityTypeName = EntityTypeName::from_str("Album").unwrap();
+    /// # let r = EntityUid::from_type_name_and_id(r_name, r_eid);
+    /// let request: Request = Request::new(Some(p), Some(a), Some(r), context);
+    /// ```
     pub fn from_json_str(
         json: &str,
         schema: Option<(&Schema, &EntityUid)>,
@@ -3086,7 +3433,7 @@ mod schema_tests {
                 }
             }}"#
             )),
-            Err(SchemaError::ParseJson(_))
+            Err(SchemaError::Serde(_))
         );
     }
 }
@@ -3717,7 +4064,7 @@ mod schema_based_parsing_tests {
     /// Test that involves open entities
     #[test]
     #[should_panic(
-        expected = "UnsupportedSchemaFeature(\"Records and entities with additional attributes are not yet implemented.\")"
+        expected = "UnsupportedFeature(\"records and entities with additional attributes are not yet implemented\")"
     )]
     fn open_entities() {
         let schema = Schema::from_str(
@@ -3785,7 +4132,7 @@ mod schema_based_parsing_tests {
     #[test]
     fn schema_sanity_check() {
         let src = "{ , .. }";
-        assert_matches!(Schema::from_str(src), Err(super::SchemaError::ParseJson(_)));
+        assert_matches!(Schema::from_str(src), Err(super::SchemaError::Serde(_)));
     }
 
     #[test]